Modified setenvironment.sh

+++ sbin/setenvironment.sh
@@ -138,19 +138,27 @@ if /usr/sbin/service lighttpd status; then
 fi
 
 # sudoers.d/lbdefault
-rm /etc/sudoers.d/lbdefaults
+if [ -L /etc/sudoers.d/lbdefaults ]; then
+    rm /etc/sudoers.d/lbdefaults
+fi
 ln -s $LBHOME/system/sudoers/lbdefaults /etc/sudoers.d/lbdefaults
 
 # profile.d/loxberry.sh
-rm /etc/profile.d/loxberry.sh
+if [ -L /etc/profile.d/loxberry.sh ]; then
+    rm /etc/profile.d/loxberry.sh
+fi
 ln -s $LBHOME/system/profile/loxberry.sh /etc/profile.d/loxberry.sh
 
 # Init Script
-rm /etc/init.d/loxberry
+if [ -L /etc/init.d/loxberry ]; then  
+   rm /etc/init.d/loxberry
+fi
 ln -s $LBHOME/sbin/loxberryinit.sh /etc/init.d/loxberry
 update-rc.d loxberry defaults
 
-rm /etc/init.d/createtmpfsfoldersinit
+if [ -L /etc/init.d/createtmpfsfoldersinit ]; then  
+   rm /etc/init.d/createtmpfsfoldersinit
+fi
 ln -s $LBHOME/sbin/createtmpfsfoldersinit.sh /etc/init.d/createtmpfsfoldersinit
 update-rc.d createtmpfsfoldersinit defaults
 
@@ -158,46 +166,60 @@ update-rc.d createtmpfsfoldersinit defaults
 if [ ! -L /etc/apache2 ]; then
 	mv /etc/apache2 /etc/apache2.old
 fi
-rm /etc/apache2
+if [ -L /etc/apache2 ]; then  
+    rm /etc/apache2
+fi
 ln -s $LBHOME/system/apache2 /etc/apache2
 
 # Lighttpd Config
 if [ ! -L /etc/lighttpd ]; then
 	mv /etc/lighttpd /etc/lighttpd.old
 fi
-rm /etc/lighttpd
+if [ -L /etc/lighttpd ]; then  
+	rm /etc/lighttpd
+fi
 ln -s $LBHOME/system/lighttpd /etc/lighttpd
 
 # Network config
 if [ ! -L /etc/network/interfaces ]; then
 	mv /etc/network/interfaces /etc/network/interfaces.old
 fi
-rm /etc/network/interfaces
+if [ -L /etc/network/interfaces ]; then  
+    rm /etc/network/interfaces
+fi
 ln -s $LBHOME/system/network/interfaces /etc/network/interfaces
 
 # Logrotate
-rm /etc/logrotate.d/loxberry
+if [ -L /etc/logrotate.d/loxberry ]; then
+    rm /etc/logrotate.d/loxberry
+fi
 ln -s $LBHOME/system/logrotate/logrotate /etc/logrotate.d/loxberry
 
 # Samba Config
 if [ ! -L /etc/samba ]; then
 	mv /etc/samba /etc/samba.old
 fi
-rm /etc/samba
+if [ -L /etc/samba ]; then
+    rm /etc/samba
+fi
 ln -s $LBHOME/system/samba /etc/samba
 
 # VSFTPd Config
 if [ ! -L /etc/vsftpd.conf ]; then
 	mv /etc/vsftpd.conf /etc/vsftpd.conf.old
 fi
-rm /etc/vsftpd.conf
+if [ -L /etc/vsftpd.conf ]; then
+    rm /etc/vsftpd.conf
+fi
 ln -s $LBHOME/system/vsftpd/vsftpd.conf /etc/vsftpd.conf
 
 # SSMTP Config
 if [ ! -L /etc/ssmtp ]; then
 	mv /etc/ssmtp /etc/ssmtp.old
 fi
-rm /etc/ssmtp
+if [ -L /etc/ssmtp ]; then
+    rm /etc/ssmtp
+fi
 ln -s $LBHOME/system/ssmtp /etc/ssmtp
 
 # PHP
@@ -226,19 +248,27 @@ awk -v s="include_path=\".:$LBHOME/libs/phplib\"" '/^include_path=/{$0=s;f=1} {a
 if [ ! -L /etc/cron.d ]; then
 	mv /etc/cron.d /etc/cron.d.old
 fi
-rm /etc/cron.d
+if [ -L /etc/cron.d ]; then
+    rm /etc/cron.d
+fi
 ln -s $LBHOME/system/cron/cron.d /etc/cron.d
 
 # Group mebership
 /usr/sbin/usermod -a -G sudo,dialout,audio,gpio,tty,www-data loxberry
 
 # Skel for system logs, LB system logs and LB plugin logs
-find $LBHOME/log/skel_system/ -exec rm {} \;
-find $LBHOME/log/skel_syslog/ -exec rm {} \;
-find $LBHOME/log/skel_plugins/ -exec rm {} \;
+if [ -d $LBHOME/log/skel_system/ ]; then
+    find $LBHOME/log/skel_system/ -type f -exec rm {} \;
+fi
+if [ -d $LBHOME/log/skel_syslog/ ]; then
+    find $LBHOME/log/skel_syslog/ -type f -exec rm {} \;
+fi
+if [ -d $LBHOME/log/skel_plugins/ ]; then
+    find $LBHOME/log/skel_plugins/ -type f -exec rm {} \;
+fi
 
 # Clean apt cache
-rm -f /var/cache/apt/archives/*
+rm -rf /var/cache/apt/archives/*
 
 # Disable PrivateTmp for Apache2 on systemd
 # (also included in 1.0.2 Update script)
@@ -246,4 +276,3 @@ if [ ! -e /etc/systemd/system/apache2.service.d/privatetmp.conf ]; then
 	mkdir -p /etc/systemd/system/apache2.service.d
 	echo -e "[Service]\nPrivateTmp=no" > /etc/systemd/system/apache2.service.d/privatetmp.conf 
 fi
-
